altos/attiny: Eliminate warnings from -Wconversion
These have no effect on the generated code as they simply make
implicit conversions explicit.

+++ src/attiny/ao_arch_funcs.h
@@ -21,7 +21,7 @@
  */
 
 #define ao_spi_get_mask(reg,mask,bus,speed) do {	\
-		(reg) &= ~(mask);			\
+		(reg) &= (uint8_t) ~(mask);		\
 	} while (0)
 
 #define ao_spi_put_mask(reg,mask,bus) do {	\
